+++ pipelines/PIPELINE-FULL-STAGING/static_test.sh
@@ -20,16 +20,19 @@ fi
 flake8 src/*.py
 if [[ $? -ne 0 ]]
 then
+    echo 'Ha fallado el análisis de estilo(flake8)'
     exit 1
 fi
 bandit src/*.py
 if [[ $? -ne 0 ]]
 then
+    echo 'Ha fallado el análisis de seguridad de codigo (bandit)'
     exit 1
 fi
 # PEP8 style guidelines
-pylint src/*.py
+pylint --errors-only src/*.py
 if [[ $? -ne 0 ]]
 then
+    echo 'Ha fallado el análisis de estilo(pylint)'
     exit 1
 fi