+++ src/get.py
@@ -1,16 +1,18 @@
+"""Code related with the get call"""
 import json
-import decimalencoder
-import todoList
+import src.decimalencoder
+import src.todoList
 
 
 def get(event, context):
+    """logic for the get"""
     # create a response
-    item = todoList.get_item(event['pathParameters']['id'])
+    item = src.todoList.get_item(event['pathParameters']['id'])
     if item:
         response = {
             "statusCode": 200,
             "body": json.dumps(item,
-                               cls=decimalencoder.DecimalEncoder)
+                               cls=src.decimalencoder.DecimalEncoder)
         }
     else:
         response = {