+++ test/integration/todoApiTest.py
@@ -2,7 +2,8 @@
 import http.client
 import os
 import unittest
-from urllib.request import urlopen
+#from urllib.request import urlopen
+import urllib
 import requests
 import json